stage5/requirements.txt
# Stage 5: CLI + web UI
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-dotenv==1.0.1
python-multipart==0.0.20
